# packages/contracts/openapi/runtime.json
{
  "components": {
    "schemas": {
      "Activation": {
        "additionalProperties": false,
        "properties": {
          "expected_scope_revision": {
            "minimum": 0.0,
            "title": "Expected Scope Revision",
            "type": "integer"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Reason",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "expected_scope_revision",
          "reason"
        ],
        "title": "Activation",
        "type": "object"
      },
...
      "AsyncAccepted": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status_url"
        ],
        "title": "AsyncAccepted",
        "type": "object"
      },
...
      "Candidate": {
        "additionalProperties": false,
        "properties": {
          "basis": {
            "title": "Basis",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "system": {
            "title": "System",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "code",
          "display",
          "system",
          "version",
          "basis"
        ],
        "title": "Candidate",
        "type": "object"
      },
      "ClinicalTime": {
        "additionalProperties": false,
        "properties": {
          "anchor_fact_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Fact Revision Id"
          },
          "derivation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derivation"
          },
          "original_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Text"
          },
          "precision": {
            "enum": [
              "year",
              "month",
              "day",
              "datetime",
              "relative",
              "unknown"
            ],
            "title": "Precision",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "precision"
        ],
        "title": "ClinicalTime",
        "type": "object"
      },
      "ClinicalValue": {
        "additionalProperties": false,
        "properties": {
          "assertion": {
            "default": "affirmed",
            "enum": [
              "affirmed",
              "negated",
              "suspected"
            ],
            "title": "Assertion",
            "type": "string"
          },
          "comparator": {
            "anyOf": [
              {
                "enum": [
                  "=",
                  "<",
                  ">",
                  "<=",
                  ">="
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparator"
          },
          "experiencer": {
            "default": "patient",
            "enum": [
              "patient",
              "family",
              "other",
              "unknown"
            ],
            "title": "Experiencer",
            "type": "string"
          },
          "kind": {
            "enum": [
              "text",
              "decimal",
              "date"
            ],
            "title": "Kind",
            "type": "string"
          },
          "missing_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MissingReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "normalized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact decimal serialized as string; never float",
            "title": "Normalized"
          },
          "raw": {
            "minLength": 1,
            "title": "Raw",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "raw",
          "normalized",
          "kind"
        ],
        "title": "ClinicalValue",
        "type": "object"
      },
      "Coding": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "fact_id": {
            "format": "uuid",
            "title": "Fact Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "unmapped"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "fact_id",
          "version",
          "status",
          "reason",
          "candidates"
        ],
        "title": "Coding",
        "type": "object"
      },
...
      "Evidence": {
        "additionalProperties": false,
        "properties": {
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/BoundingBox"
            },
            "title": "Boxes",
            "type": "array"
          },
          "document_version_id": {
            "format": "uuid",
            "title": "Document Version Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "parse_artifact_id": {
            "format": "uuid",
            "title": "Parse Artifact Id",
            "type": "string"
          },
          "quote": {
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          },
          "span": {
            "$ref": "#/components/schemas/TextSpan"
          },
          "text_version": {
            "minLength": 1,
            "title": "Text Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_version_id",
          "parse_artifact_id",
          "text_version",
          "page",
          "span",
          "quote"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "ExtractionAccepted": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "job_id",
          "status_url"
        ],
        "title": "ExtractionAccepted",
        "type": "object"
      },
      "ExtractionConfigView": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "ocr_provider": {
            "title": "Ocr Provider",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "terminology_version": {
            "title": "Terminology Version",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "enabled",
          "ocr_provider",
          "terminology_version"
        ],
        "title": "ExtractionConfigView",
        "type": "object"
      },
      "ExtractionCreate": {
        "additionalProperties": false,
        "properties": {
          "parse_artifact_id": {
            "format": "uuid",
            "title": "Parse Artifact Id",
            "type": "string"
          },
          "template_version": {
            "enum": [
              "outpatient-1.0.0",
              "laboratory-1.0.0"
            ],
            "title": "Template Version",
            "type": "string"
          }
        },
        "required": [
          "template_version",
          "parse_artifact_id"
        ],
        "title": "ExtractionCreate",
        "type": "object"
      },
      "ExtractionView": {
        "additionalProperties": false,
        "properties": {
          "attempt_measurements": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Attempt Measurements",
            "type": "array"
          },
          "codings": {
            "items": {
              "$ref": "#/components/schemas/Coding"
            },
            "title": "Codings",
            "type": "array"
          },
          "configuration": {
            "additionalProperties": true,
            "title": "Configuration",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_version_id": {
            "format": "uuid",
            "title": "Document Version Id",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactRevision"
            },
            "title": "Facts",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_digest": {
            "title": "Input Digest",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "title": "Issues",
            "type": "array"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "parse_artifact_id": {
            "format": "uuid",
            "title": "Parse Artifact Id",
            "type": "string"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/Relation"
            },
            "title": "Relations",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          }
        },
        "required": [
          "id",
          "job_id",
          "document_version_id",
          "parse_artifact_id",
          "configuration",
          "input_digest",
          "created_at",
          "status",
          "error",
          "duration_ms",
          "attempt_measurements",
          "usage",
          "facts",
          "relations",
          "issues",
          "codings"
        ],
        "title": "ExtractionView",
        "type": "object"
      },
      "FactRevision": {
        "additionalProperties": false,
        "properties": {
          "entity_group_id": {
            "format": "uuid",
            "title": "Entity Group Id",
            "type": "string"
          },
          "event_time": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClinicalTime"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "minItems": 1,
            "title": "Evidence",
            "type": "array"
          },
          "excluded": {
            "default": false,
            "title": "Excluded",
            "type": "boolean"
          },
          "fact_id": {
            "format": "uuid",
            "title": "Fact Id",
            "type": "string"
          },
          "field_path": {
            "minLength": 1,
            "title": "Field Path",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/ClinicalValue"
          }
        },
        "required": [
          "fact_id",
          "revision_id",
          "revision",
          "entity_group_id",
          "field_path",
          "value",
          "evidence"
        ],
        "title": "FactRevision",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Issue": {
        "additionalProperties": false,
        "properties": {
          "can_accept_unknown": {
            "title": "Can Accept Unknown",
            "type": "boolean"
          },
          "fact_revision_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Fact Revision Ids",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "rule": {
            "title": "Rule",
            "type": "string"
          },
          "severity": {
            "enum": [
              "blocking",
              "warning",
              "info"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "const": "open",
            "default": "open",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "rule",
          "severity",
          "fact_revision_ids",
          "message",
          "can_accept_unknown"
        ],
        "title": "Issue",
        "type": "object"
      },
...
      "Relation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "target_id",
          "kind"
        ],
        "title": "Relation",
        "type": "object"
      },
      "ReviewSetView": {
        "additionalProperties": false,
        "properties": {
          "encounter_id": {
            "format": "uuid",
            "title": "Encounter Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/Issue"
            },
            "title": "Issues",
            "type": "array"
          },
          "members": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Members",
            "type": "array"
          },
          "scope_revision": {
            "title": "Scope Revision",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "encounter_id",
          "scope_revision",
          "status",
          "members",
          "issues"
        ],
        "title": "ReviewSetView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "临床数据结构化平台 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/documents/{document_id}/active-run": {
      "post": {
        "operationId": "activateExtraction",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Activation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate"
      }
    },
...
    "/api/v1/projects/{project_id}/documents/{document_id}/extractions": {
      "get": {
        "operationId": "listExtractions",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExtractionView"
                  },
                  "title": "Response Listextractions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Extractions"
      },
      "post": {
        "operationId": "createExtraction",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Extraction"
      }
    },
    "/api/v1/projects/{project_id}/documents/{document_id}/ocr": {
      "post": {
        "operationId": "startOCR",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Ocr"
      }
    },
...
    "/api/v1/projects/{project_id}/encounters/{encounter_id}/review-set": {
      "get": {
        "operationId": "getEncounterReviewSet",
        "parameters": [
          {
            "in": "path",
            "name": "encounter_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Encounter Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review"
      }
    },
    "/api/v1/projects/{project_id}/extraction-config": {
      "get": {
        "operationId": "getExtractionConfig",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionConfigView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Config"
      }
    },
    "/api/v1/projects/{project_id}/extractions/{run_id}": {
      "get": {
        "operationId": "getExtraction",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Extraction"
      }
    },
...
    "/api/v1/projects/{project_id}/terminology/candidates": {
      "get": {
        "operationId": "getTerminologyCandidates",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Candidate"
                  },
                  "title": "Response Getterminologycandidates",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Terminology"
      }
    },
...
  }
}